Refactored /api/users/ to be much simpler.
